backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Redis and vector search
redis>=5.0.0
redisvl>=0.2.0

# Data processing
pandas>=2.0.0

# ML/Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# NLP
nltk>=3.8.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0

# LLM
openai>=1.0.0

